TSSOP14 Footprint fix from last commit. Ran DRC
In the previous commit, I set the origin reference of the TSSOP14 part to "center" and forgot to change it back to its original position. That is fixed now. I also ran DRC and removed some antenna traces (they were just little unconnected stubs) and relaxed some of the rules regarding silkscreen clearances. Not all the rules are resolved, but what's left is probably going to end up being waived, but I want to look through them all anyway. Mostly silk to silk clearance issues in the TC part of the board, but for most of them it's fine if some of the silk gets mushed together.