Grounded mounting holes. Fixed missing connections
The mounting holes were grounded because the screws going through them will be going directly into an aluminum box that will shield them from the radio beacon inside the rocket, and that shield needs to be grounded.

The missing connections I found were two ground pins on the microcontroller and one ground pin on one of the thermocouple ICs that were not connected to ground. Fixed that now.