Implemented David and Tyler's final revisions
Added 45 uF of bulk storage capacitors to LM317s and 15 uF to each valve driver IC to keep the 12V line from dipping when valves draw lots of power. 

Added pull-down resistors to each used enable line for the valve drivers to prevent the line from going high during MCU booting. 

Increased by 10x the resistor values of R6 and R7 (the divider on pressure transducer output) to reduce load on the transducers and be more power efficient. 

Added connector for UART. 

Increased copper pad area and added thermal vias to the LM317 ICs. Complemented this by making the power rails transmitted by copper polygons, for the most part. Also added bottom copper layers to LM317 and bridge driver ICs for additional heat dissipation. 

Switched the order of the ground and power layers so that the ground layer is now closer to the top copper signal layer. 

Added grounded pads in between each pair of thermocouple connectors to allow for the TC wire shielding to be grounded.