Rewrote the force code and messed up the old code
The new force code seems to work now. In the process of trying to get the old force code to work I think I messed it up further, but I'm comitting it anyway b/c I'm scared. This is a private repo anyway so there's no chance of an employer seeing my crappy commit style, right?